Use bindMemory() instead of withMemoryRebound()
Only use this method to rebind the pointer’s memory to a type with the same size and stride as the currently bound Pointee type. To bind a region of memory to a type that is a different size, convert the pointer to a raw pointer and use the bindMemory(to:capacity:) method.

https://developer.apple.com/documentation/swift/unsafepointer/2430863-withmemoryrebound