Allow multiple instances of WasmInterpreter to have imported functions (#14)
This commit fixes a bug in which deinitializing an instance of `WasmInterpreter` would remove all cached implementations for functions imported by any `WasmInterpreter` instance, not just the one being deinitialized.

Now, we generate a unique identifier for each instance of `WasmInterpreter`, and we use that identifier as the key to its cached imported functions.

This pull request also increases to 10 the maximum number of Wasm functions that can share the same implementation in the same module. Previously, the limit was 3.